+++ lib/packages/wikibase/data-model/src/Term/AliasGroup.php
@@ -18,7 +18,14 @@ use InvalidArgumentException;
  */
 class AliasGroup implements Comparable, Countable {
 
+	/**
+	 * @var string
+	 */
 	private $languageCode;
+
+	/**
+	 * @var string[]
+	 */
 	private $aliases;
 
 	/**